러닝 기록 db에 저장 성공

+++ src/main/java/com/runaway/project/running/dto/RunningDto.java
@@ -16,7 +16,7 @@ public class RunningDto {
 
     private Long idx;
 
-    private int distance;
+    private double distance;
 
     private String runningTime;
